+++ NEWS.md
@@ -1,3 +1,8 @@
+# rollama 0.1.0
+
+* adds function `check_model_installed`
+* changes default model to llama3
+
 # rollama 0.0.3
 
 * add option to query several models at once